Move consensus rules framework to base
