+++ src/Stratis.Bitcoin.IntegrationTests.Common/Extensions/CoreNodeExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using NBitcoin;
+using Stratis.Bitcoin.Consensus.Rules;
 using Stratis.Bitcoin.Features.Consensus.Rules;
 using Stratis.Bitcoin.Features.Consensus.Rules.CommonRules;
 using Stratis.Bitcoin.Features.Wallet;
